tweaked bats bats now look around fixed map2 fishes going out of the map

+++ Game/src/game/entity/living/enemy/EntityEnemy.java
@@ -123,7 +123,7 @@ public class EntityEnemy extends EntityLiving {
 		flinching = true;
 	}
 	
-	public int getAttackDamage(){
+	public float getAttackDamage(){
 		return 1;
 	}
